feat(configuration): init configuration
moved from old config package
