--- packages/config/src/prettier/config.ts
@@ -1,9 +1,0 @@
-import { PrettierConfigModel } from "./PrettierConfigModel";
-
-export const prettierConfig = {
-    singleQuote: false,
-    trailingComma: "all",
-    tabWidth: 4,
-    printWidth: 120,
-    endOfLine: "auto",
-} as PrettierConfigModel;
