+++ packages/configuration/.eslintrc.json
