+++ packages/configuration/README.md
