+++ packages/configuration/jest.config.ts
@@ -11,5 +11,5 @@ export default {
     '^.+\\.[tj]s$': 'ts-jest',
   },
   moduleFileExtensions: ['ts', 'js', 'html'],
-  coverageDirectory: '../../coverage/packages/config',
+  coverageDirectory: '../../coverage/packages/configuration',
 };
