+++ packages/configuration/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@weichwarenprojekt/config",
+  "name": "@weichwarenprojekt/configuration",
   "version": "0.0.1",
   "type": "commonjs"
 }
