+++ packages/configuration/project.json
@@ -1,23 +1,23 @@
 {
-  "name": "@weichwarenprojekt/config",
   "$schema": "../../node_modules/nx/schemas/project-schema.json",
-  "sourceRoot": "packages/config/src",
+  "name": "@weichwarenprojekt/configuration",
+  "sourceRoot": "packages/configuration/src",
   "projectType": "library",
   "targets": {
     "build": {
       "executor": "@nrwl/js:tsc",
       "outputs": ["{options.outputPath}"],
       "options": {
-        "outputPath": "dist/packages/config",
-        "main": "packages/config/src/index.ts",
-        "tsConfig": "packages/config/tsconfig.lib.json",
-        "assets": ["packages/config/*.md"]
+        "outputPath": "dist/packages/configuration",
+        "main": "packages/configuration/src/index.ts",
+        "tsConfig": "packages/configuration/tsconfig.lib.json",
+        "assets": ["packages/configuration/*.md"]
       }
     },
     "publish": {
       "executor": "nx:run-commands",
       "options": {
-        "command": "node tools/scripts/publish.mjs config",
+        "command": "node tools/scripts/publish.mjs configuration",
         "parallel": false
       },
       "dependsOn": ["build"]
@@ -26,14 +26,14 @@
       "executor": "@nrwl/linter:eslint",
       "outputs": ["{options.outputFile}"],
       "options": {
-        "lintFilePatterns": ["packages/config/**/*.ts"]
+        "lintFilePatterns": ["packages/configuration/**/*.ts"]
       }
     },
     "test": {
       "executor": "@nrwl/jest:jest",
       "outputs": ["{workspaceRoot}/coverage/{projectRoot}"],
       "options": {
-        "jestConfig": "packages/config/jest.config.ts",
+        "jestConfig": "packages/configuration/jest.config.ts",
         "passWithNoTests": true
       }
     },
