+++ packages/configuration/tsconfig.json
