+++ packages/configuration/tsconfig.lib.json
